ErdblickRenderer: Restructured project.
