+++ libs/renderer/src/FeatureLayerStyle.cpp
@@ -1,9 +1,10 @@
 #include <iostream>
 
-#include "include/FeatureLayerStyle.h"
 #include "yaml-cpp/yaml.h"
 #include "simfil/simfil.h"
 #include "simfil/model/nodes.h"
+
+#include "FeatureLayerStyle.h"
 
 FeatureLayerStyle::FeatureLayerStyle(SharedUint8Array& yamlArray)
 {
