+++ libs/renderer/src/duckfile.c
@@ -1,4 +1,4 @@
-#include <stdlib.h>
+#include "../../../ci/emsdk/upstream/emscripten/cache/sysroot/include/compat/stdlib.h"
 const char duckfile[] = "\
 glTF\002\000\000\000\244\326\001\000\100\b\000\000JSON{\"asset\":{\"generator\":\"COLLADA2GLTF\",\"version\":\"2.0\"},\"scene\":0,\"scenes\":[{\"nodes\":[0]}],\"nodes\":[{\"children\":[2,1],\"matrix\":[0.009999999776482582,0.0,0.0,0.0,0.0,0.009999999776482582,0.0,0.0,0.0,0.0,0.009999999776482582,0.0,0.0,0.0,0.0,1.0]},{\"matrix\":[-0.7289686799049377,0.0,-0.6845470666885376,0.0,-0.42520490288734438,0.7836934328079224,0.4527972936630249,0.0,0.5364750623703003,0.6211478114128113,-0.571287989616394,0.0,400.1130065917969,463.2640075683594,-431.0780334472656,1.0],\"camera\":0},{\"mesh\":0}],\"cameras\":[{\"perspective\":{\"aspectRatio\":1.5,\"yfov\":0.6605925559997559,\"zfar\":10000.0,\"znear\":1.0},\"type\":\"perspective\"}],\"meshes\":[{\"primitives\":[{\"attributes\":{\"NORMAL\":1,\"POSITION\":2,\"TEXCOORD_0\":3},\"indices\":0,\"mode\":4,\"material\":0}],\"name\":\"LOD3spShape\"}],\"accessors\":[{\"bufferView\":0,\"byteOffset\":0,\"componentType\":5123,\"count\":12636,\"max\":[2398],\"min\":[0],\"type\":\"SCALAR\"},{\"bufferView\":1,\"byteOffset\":0,\"componentType\":5126,\"count\":2399,\"max\":[0.9995989799499512,0.999580979347229,0.9984359741210938],\"min\":[-0.9990839958190918,-1.0,-0.9998319745063782],\"type\":\"VEC3\"},{\"bufferView\":1,\"byteOffset\":28788,\"componentType\":5126,\"count\":2399,\"max\":[96.17990112304688,163.97000122070313,53.92519760131836],\"min\":[-69.29850006103516,9.929369926452637,-61.32819747924805],\"type\":\"VEC3\"},{\"bufferView\":2,\"byteOffset\":0,\"componentType\":5126,\"count\":2399,\"max\":[0.9833459854125977,0.9800369739532471],\"min\":[0.026409000158309938,0.01996302604675293],\"type\":\"VEC2\"}],\"materials\":[{\"pbrMetallicRoughness\":{\"baseColorTexture\":{\"index\":0},\"metallicFactor\":0.0},\"emissiveFactor\":[0.0,0.0,0.0],\"name\":\"blinn3-fx\"}],\"textures\":[{\"sampler\":0,\"source\":0}],\"images\":[{\"bufferView\":3,\"mimeType\":\"image/png\"}],\"samplers\":[{\"magFilter\":9729,\"minFilter\":9986,\"wrapS\":10497,\"wrapT\":10497}],\"bufferViews\":[{\"buffer\":0,\"byteOffset\":76768,\"byteLength\":25272,\"target\":34963},{\"buffer\":0,\"byteOffset\":0,\"byteLength\":57576,\"byteStride\":12,\"target\":34962},{\"buffer\":0,\"byteOffset\":57576,\"byteLength\":19192,\"byteStride\":8,\"target\":34962},{\"buffer\":0,\"byteOffset\":102040,\"byteLength\":16302}],\"buffers\":[{\"byteLength\":118342}]}  H\316\001\000BIN\0009\270D\276\352\077o\277\217R\231>\311T\201\275\024^~\277+L\277={\203\357\275+\333k\277\230\333\275>%\312\036\275\2235~\277\313h\344=\344\236\256\275\000r~\277\267z\216=\0324\204\276\345+q\277\0052[>(H\234\276\204\271q\277i\337\374=\365\367\322\275vn~\277p \044=\203\370\340\275\376b~\277n\330\266<\"\306\243\276\356\tr\277\t\337{=w\272s>T7O\277\rl\t\077\305\212\312>\021QT\277\020\002\312>&\337\244>\3441'\277\270u/\0774\017\f\077]p*\277X\342\001\077\376\324h>\347pq\277\370Rx>\245\335\030>\226\am\277\352\255\261>\240\377\236\276;\027R\2775\225\365>\377y:\276\214JJ\277\201\315\025\077W\044\326\2764eW\277\247;\257>k\202\320\276u\004,\277d[\036\077\372}\v\277F\0776\277\313\326\342>\364\376o\276\336;\036\277\276\026\100\077\304\320\362\276){[\277\020\350L>|\355\361\276\000\033`\277y\261\320=\227\343\031\277)%\100\277\242}\214>O\352\027\277\366\232J\277\304\"\026>\206\221\376>'kX\277\226\351G>\334\203\220>~\031t\277Q/\330=\212 .\077I\273-\277\227\343\215>\375\025\312>\245\022\002\277]\372C\077\032\334&\077\351\357\001\277\356A\020\077\350j\363>\020\347\301\276\033HK\077\000\215:\077xz\265\276\004\001\026\077\376\264\361\276\225\t\a\277\205\3214\077Mg\217\276\375\372\351\276\357\036X\077VD!\277\006\360\026\277\307h\001\077\001\211\006\277\201=\316\276\n\
 \327\077\077f20\277\271\210\367\276\352v\n\
