+++ static/index.html
@@ -10,7 +10,7 @@
 
 <div id="demoScene"></div>
 
-<script src="./FeatureLayerRenderer.js"></script>
+<script src="./libs/renderer/ErdblickRenderer.js"></script>
 
 <!-- THREE.js sources -->
 <script type="module" src="./deps/three.js"></script>
